requirements.txt
yt-dlp>=2024.4.9
python-telegram-bot
requests